patient-monitoring/terraform/script/body-temp/requirements.txt
nltk~=3.8.1
kafka-python